cloud/backend/requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary
pillow
djangorestframework-simplejwt